backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.127.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-multipart>=0.0.21",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
